+++ src/Transformer/ObjectToObjectMetadata/ObjectToObjectMetadata.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * that was distributed with this source code.
  */
 
-namespace Rekalogika\Mapper\Transformer\ObjectToObjectMetadata\Contracts;
+namespace Rekalogika\Mapper\Transformer\ObjectToObjectMetadata;
 
 final class ObjectToObjectMetadata
 {