+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * feat: Data collector.
 * feat: Web profiler bundle integration.
 * fix: Fix web profiler HTML.
+* refactor: Namespace for `ArrayLikeMetadata` & `ObjectToObjectMetadata`.
 
 ## 0.6.1